🧹 Model Cleanup
