+++ app/models/__init__.py
@@ -1,9 +1,3 @@
-from .auxiliary_key_pair import *
-from .auxiliary_public_key import *
 from .base import *
-from .election_key_pair import *
-from .election_key_pair_request import *
 from .guardian import *
-from .guardian_backup import *
-from .guardian_backup_request import *
-from .guardian_request import *
+from .key import *
