--- app/models/auxiliary_key_pair.py
@@ -1,6 +1,0 @@
-from .base import Base
-
-
-class AuxiliaryKeyPair(Base):
-    secret_key: str
-    public_key: str
