--- app/models/auxiliary_public_key.py
@@ -1,7 +1,0 @@
-from .base import Base
-
-
-class AuxiliaryPublicKey(Base):
-    owner_id: str
-    sequence_order: int
-    key: str
