--- app/models/election_key_pair.py
@@ -1,10 +1,0 @@
-from typing import Any
-
-from .base import Base
-
-
-class ElectionKeyPair(Base):
-    secret_key: str
-    public_key: str
-    proof: Any
-    polynomial: Any
