--- app/models/election_key_pair_request.py
@@ -1,8 +1,0 @@
-from typing import Optional
-
-from .base import Base
-
-
-class ElectionKeyPairRequest(Base):
-    quorum: int
-    nonce: Optional[str] = None
