--- app/models/guardian_request.py
@@ -1,11 +1,0 @@
-from typing import Optional
-
-from .base import Base
-
-
-class GuardianRequest(Base):
-    id: str
-    sequence_order: int
-    number_of_guardians: int
-    quorum: int
-    nonce: Optional[str] = None
